requirements.txt
google-genai>=0.2.0
pypdf2>=3.0.0
pillow>=10.0.0
annoy>=1.17.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
customtkinter>=5.2.0
